move config tests into main_config_test.go
